Add namespaces promise to namespace list Add namespace promise to namespace list and navigation
