+++ src/lib/components/namespace-list.svelte
@@ -1,19 +1,79 @@
 <script lang="ts">
   import { page } from '$app/stores';
+  import { createEventDispatcher } from 'svelte';
+  import { onMount } from 'svelte';
 
   import Icon from '$holocene/icon/icon.svelte';
-
-  import { onMount } from 'svelte';
   import EmptyState from '$lib/holocene/empty-state.svelte';
-  import { createEventDispatcher } from 'svelte';
+  import Loading from '$lib/holocene/loading.svelte';
 
-  export let getNamespaceList: () => Promise<NamespaceItem[]> = null;
+  import { lastUsedNamespace } from '$lib/stores/namespaces';
+  import { workflowSorts, workflowFilters } from '$lib/stores/filters';
+  import { goto } from '$app/navigation';
+  import {
+    routeForWorkflows,
+    routeForSchedules,
+  } from '$lib/utilities/route-for';
+  import { fetchNamespaces } from '$lib/services/namespaces-service';
 
-  let namespaceList = null;
+  import type { DescribeNamespaceResponse as Namespace } from '$types';
+  import type { ListNamespacesResponse } from '$types';
+
   let searchField: HTMLInputElement = null;
+  const {
+    showTemporalSystemNamespace,
+    runtimeEnvironment: { isCloud },
+  } = $page.stuff.settings;
+
+  const namespacesPromise: Promise<ListNamespacesResponse> = fetchNamespaces(
+    $page.stuff.settings,
+  );
+
+  function getCurrentHref(namespace: string) {
+    const onSchedulesPage = $page.url.pathname.endsWith('schedules');
+    const href = onSchedulesPage
+      ? routeForSchedules({ namespace })
+      : routeForWorkflows({ namespace });
+    return href;
+  }
+
+  function getNamespaceList(namespaces: Namespace[]) {
+    const filteredNamespaces: string[] = namespaces
+      .map((namespace: Namespace) => namespace?.namespaceInfo?.name)
+      .filter(
+        (namespace: string) =>
+          showTemporalSystemNamespace || namespace !== 'temporal-system',
+      );
+    let namespaceList = filteredNamespaces.map((namespace: string) => {
+      return {
+        namespace,
+        href: (namespace: string) => getCurrentHref(namespace),
+        onClick: (namespace: string) => {
+          $lastUsedNamespace = namespace;
+          $workflowFilters = [];
+          $workflowSorts = [];
+          goto(getCurrentHref(namespace));
+        },
+      };
+    });
+
+    // To show single namespace on cloud
+    if (isCloud && $page.params.namespace && !filteredNamespaces.length) {
+      namespaceList.push({
+        namespace: $page.params.namespace,
+        href: (namespace: string) => routeForWorkflows({ namespace }),
+        onClick: (namespace: string) => {
+          $lastUsedNamespace = $page.params.namespace;
+          $workflowFilters = [];
+          $workflowSorts = [];
+          goto(routeForWorkflows({ namespace }));
+        },
+      });
+    }
+    return namespaceList;
+  }
 
   onMount(() => {
-    namespaceList = getNamespaceList();
     searchField.focus();
   });
   const dispatch = createEventDispatcher();
@@ -61,11 +121,12 @@
 </div>
 
 <ul data-cy="namespace-list">
-  {#await namespaceList}
-    Loading ...
+  {#await namespacesPromise}
+    <Loading />
   {:then namespacesResult}
-    {#if namespacesResult}
-      {#each namespacesResult.filter( ({ namespace }) => namespace.includes(searchValue), ) as namespace}
+    {@const namespaceList = getNamespaceList(namespacesResult.namespaces)}
+    {#if namespaceList.length > 0}
+      {#each namespaceList.filter( ({ namespace }) => namespace.includes(searchValue), ) as namespace}
         <li
           class="first:rounded-t-md first:border-t last:rounded-b-md border-b border-l border-r p-3 flex border-collapse gap-2 hover:bg-gray-50 cursor-pointer"
           on:click={() => namespace?.onClick(namespace.namespace)}
